Testing sizes.
I have combined all the separate parts that I would like to print. 
I had added in the R-pi, power board thing and chassis to be able to visualize it together.. the parts that I am wanting to print are the two front braces that have the IR mounts on them (the IR mounts are angled at 45.), the camera mount and two rear braces. 
Have a look and let me know what you think could do with improving. 
Have based this on what we discussed during the LABS. 
Alex.